+++ R/pbtypes.R
@@ -1,7 +1,7 @@
 pbtypes <-
 function()
 {
-    TYPES <- c("timer", "txt", "tk", "none")
+    TYPES <- c("timer", "txt", "tk", "none", "fct")
     if (.Platform$OS.type == "windows")
         c(TYPES, "win") else TYPES
 }