+++ R/unix/getpb.R
@@ -4,6 +4,7 @@ function(pb)
     if (dopb()) {
         progress.bar <- getOption("pboptions")$type
         rval <- switch(progress.bar,
+            fct = getFctProgressBar(pb),
             timer = getTxtProgressBar(pb),
             txt = getTxtProgressBar(pb),
             tk = tcltk::getTkProgressBar(pb))