+++ R/zzz.R
@@ -11,7 +11,9 @@
         nout = 100L,
         min_time = 0,
         use_lb = FALSE,
-        file = "")
+        file = "",
+        fct = function(txt) cat(txt),
+        label_name = "")
     optsx <- getOption("pboptions")
     if (!is.null(optsx)) {
         for (i in intersect(names(opts), names(optsx)))